Antisamy Stream WorkAround
Creating a ByteArrayInputStream in the HTMLValidationRule to circumvent
the undesired stream close event when processing a malformed xml
configuraiton while antisamy schema validation is disabled.

Moving the property test to an isolated file.
