+++ src/test/java/org/owasp/esapi/reference/validation/HTMLValidationRuleClasspathTest.java
@@ -22,6 +22,7 @@ import static org.junit.Assert.fail;
 
 import org.junit.After;
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
@@ -50,10 +51,6 @@ import org.owasp.validator.html.PolicyException;
  * the cleansed (sanitizied) output when certain unsafe input is encountered.
  */
 public class HTMLValidationRuleClasspathTest {
-    /**
-     * Property specified by the Antisamy project which may be used to disable schema validation on policy files.
-     */
-    private static final String ANTISAMY_PROJECT_PROP_SCHEMA_VALIDATION = "owasp.validator.validateschema";
     /** The intentionally non-compliant AntiSamy policy file. We don't intend to
      * actually <i>use</i> it for anything.
      */
@@ -93,7 +90,6 @@ public class HTMLValidationRuleClasspathTest {
     @After
     public void tearDown() throws Exception {
         ESAPI.override(null);
-        System.clearProperty(ANTISAMY_PROJECT_PROP_SCHEMA_VALIDATION);
     }
 
 	@Before
@@ -102,6 +98,7 @@ public class HTMLValidationRuleClasspathTest {
 			new ConfOverride( ESAPI.securityConfiguration(), "throw", ANTISAMY_POLICY_FILE_NONSTANDARD_LOCATION )
 		);
     }
+
 	
 	@Test
     public void checkPolicyExceptionWithBadConfig() throws Exception {
@@ -110,13 +107,6 @@ public class HTMLValidationRuleClasspathTest {
         HTMLValidationRule.loadAntisamyPolicy(INVALID_ANTISAMY_POLICY_FILE);
     }
 	
-	@Test
-    public void checkAntisamySystemPropertyWorksAsAdvertised() throws Exception {
-        ESAPI.override(null);
-        System.setProperty( ANTISAMY_PROJECT_PROP_SCHEMA_VALIDATION, "false" ); 
-        HTMLValidationRule.loadAntisamyPolicy(INVALID_ANTISAMY_POLICY_FILE);
-    }
-
     @Test
     public void testGetValid() throws Exception {
         System.out.println("getValidCP");
